Add a regression test script
This just builds riscv-tools and then runs the C++ emulator tests for
the default coniguration.  The general idea here is to put most of the
regression logic in this script, both so users can run it and so the
various regression systems don't end up with all this logic baked into
them.