Add asynchronous FIFOs as an option for clock domain crossings.
This extends AsyncDecoupledTo and AsyncDecoupledFrom to instantiate
async fifos if requested by setting depth > 0.  The modules use
the standard technique of gray-coding the read and write pointers
for synchronization, and are ported from verilog implementations
previously used for this purpose.

Note that only FIFO depths which are powers of 2 are supported.