Integrate the firrtl call into Generator
This is a bit ugly: it still uses the odd command-line argument passing
mechanism that firrtl uses, and it still serializes the firrtl out to a
file before reading it back in again, but at least firrtl and chisel are
now run in the same process.

I did this because I want to add some firrtl passes but I don't want to
add any more badness to the existing firrtl driver code.