Split the TestHarness into its own Verilog file
When I run things through the CAD tools I only want to synthesize the
top level but I still want to be able to simulate the test harness.
This patch facilitates this use case by splitting the test harness out
into a seperate Verilog file when building Rocket Chip.  This is
implemented by running the FIRRTL compiler twice with two different sets
of passes: one swaps out ExampleTop for a black box while the other
emits only ExampleTop and its children.

Note that this doesn't do the prefixing that the memserdes did, so it'll
break if the test harness uses modules that also end up in the top.