fixing up memory system
There was a bit of confusion about the switcher. Its inputs are the
coreplex memory channels, not the banks. I've changed the argument names
to avoid further confusion.

Also for multichannel configurations, we need a channel unmapper in
between to restore the addresses.

Corrected a few TileLink parameters here and there.
The switcher needs a different one with a larger client_xact_id since
it's mixing down all the channels into one.
The bidirectional LBWIF SERDES is weird because for one end it connects
to the MMIO network, whereas for the other it connects to memory
network. To make sure everything works out, we give it whichever one has
the largest client_xact_id size.

I suspect even this isn't fully correct because it can also be routed to
the edge slave. We'll have to rethink this when we reorganize the SCR
routing.