Only include/build rom.v if the config needs it
This change conditionally updates what verilog is needed
to build a simulator based on *.rom.conf
In order to get this working we need to re-excute make after
chisel has been run, this is done whenever an included makefile
is changed. Before now we have only needed this re-execution for
the list of tests to run but we now need it to know how to build
the simulator.
We also exclude this inclusion for dryruns.