Increase lrscCycles to work without branch prediction
Since an LR/SC sequence can have taken branches/jumps within the loop,
Rockets with no branch prediction can fail to meet the forward-progress
guarantee.  Increase the timeout to handle the case that all instructions
hit in the I$ but up to 14 of them are mispredicted branches.
(14 is 16 minus the LR and the SC instructions themselves.)