Move RoCC logic into RocketTile, add HasFpuOpt
To avoid problems with inheritance/trait mix-in when building RoCC-enabled
configurations, this adds a new trait, HasFpuOpt, that declares the
optional FPU so that this exists before use in the implementation of lazy
RoCCs.

RoCC connection logic to the core is moved inside the implementation of
RocketTile so that the core is declared before it is used.

The RoCC cmd/resp routers are changed to be externally visible as
Option[T].

The type variable on HasLazyRoCCModule is removed as this was never used.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>