Add HierarchicalNode.
Diplomatic node assignments can be chained together. This has many
use-cases, including many simple adapters being composed to make a
complex adapter.

It would be useful sometimes to treat a chain of connected nodes as a
single diplomatic node. For example, a crossbar with a fixed
parameterization could be implemented as

  AdapterA := FixedXBar := AdapterB

Ideally, this would chain would be encapsulated as a single node XBar so
you can do both

  XBar := rhs
  lhs := XBar

However, in the current version of diplomacy you would need two nodes
for the XBar: one for when it is the lhs of assignments and another when
it is the rhs.

This commit adds a HierarchicalNode that encapsulates other nodes. A
HierarchicalNode passes connect operations to the appropriate side of a
chain of diplmatic nodes.