Adding no exclude fire() to DecoupledHelper
Adding a no-argument fire() to DecoupledHelper that doesn't exclude anything. The previous way of doing this, which was fire(false.B), was broken by the previous change of requiring all excluded to be included, and forced a further hack of adding a dummy value. Since excluding false.B was a hack in itself, this avoids the issue entirely by providing a straightforward and safe way of excluding nothing.