Change target of reg desc annotation to IO.
Currently, the target of the reg desc annotation is the module with the
register mapping. However, a module could have multiple memory
interfaces with their own distinct regmaps, and downstream tools
(especially for unit verification) may make a distinction between the
interfaces.

Currently, the target of the annotation is the module so you can't tell
which regmap would go with a particular memory interface. This commit
resolves this issue by making the target of the annotation the bundle
corresponding to the memory interface.