Bump hardfloat, clean up build. (#1865)
Hardfloat used to require Chisel 2 for testing, which is only published
for Scala 2.11. Rocket uses Scala 2.12, which meant the build system
needed a workaround to make hardfloat use Scala 2.12 when being used as
a part of rocket.

Hardfloat's tests have been updated to use Chisel 3, so the workaround
is no longer needed. This commit bumps the hardfloat submodule and
cleans up build.sbt to use Scala 2.12 only and Makefrag to not use
++2.12.4.