Remove root logical tree node.
The notion of a global "root" logical tree node causes issues with putting
another layer of logical tree nodes above the previous "root". If a node
attaches itself to the global root, then it no longer becomes possible to insert
a different node in between the root and the previous node. In order to ensure
that it is always possible to add another layer above, we remove the concept of
a static, global "root" and require the next layer to attach the child nodes.