Add RoCC BlackBoxExample, vsrc/RoccBlackBox.v (#1937)
This adds a fourth RoCC example, BlackBoxExample, that uses a Verilog
resource defined in vsrc/RoccBlackBox.v to provide the underlying RoCC
module implementation. The included vsrc/RoccBlackBox.v accumulates
rs1 and rs2 into a persistent accumulator. Rocket Chip parameters are
passed via Verilog parameters to define IO widths.

The intended use case here is for a developer that wants to build
RoCCs in Verilog and integrate with Rocket Chip. The example does not
deal with Diplomatic IO (TL). An example including this is complicated
by the fact that Verilog doesn't handle optional/parametric IO.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>