D$: guarantee that no-alloc accesses are ordered even if aliased
Currently, non-allocating accesses can be reordered with allocating
accesses if they are to different virtual synonyms.  Add hardware
to detect and prevent this case.

If there's no possibility of aliasing or non-allocating accesses,
there's no additional HW cost.

I'm not entirely happy with the approach (it adds a lot of fanout to
what is sometimes a critical timing path, so at minimum it's leaky).
Might revisit later.