Fix a rare D$ ECC bug
This bug can manifest with a store, load, store to same word as the first
one, then load, as follows. The first store succeeds, but is still in the
pending store buffer.  Some unrelated load occurs.  Then a store to the
same word as the first store occurs, this time detecting an ECC error.
Finally, some unrelated load occurs.  (The loads are mostly irrelevant,
but the pipeline needs to stay busy for the bug to manifest.)

Since the bug can only manifest with two nearby stores to the same word,
one detecting an error and one not detecting an error, this bug is not
especially likely to occur in practice.  In particular, if the error grew
before the sequence began, the bug would not manifest.