Improve DecodeLogic runtime
For cases where none of the clauses (nor the default) is X, skip the
optimization problem altogether.

It turns out this is both a common and problematic case.  For a design I'm
working on, it reduced the runtime to about 10s from [whenver I Ctrl-C'd the
thing after several minutes].