rm *-dirty test if mem not SV39 superpage aligned
Remove the rv(32|64)si-[pv]-dirty test from the generated test suite
if an SoC's memory base address is not aligned to an SV39 superpage.
This is a hard requirement to even compile that test so there's no
point in including it in the generated test suite.

In effect, this removes the test from both of the following Makefile
targets:

  - rv64si-p-asm-tests
  - regression-tests

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>