TLToAHB: add a register to retain hrdata during an error response
It has been deemed desirable to output the data value of a read response
which fails, hopefully reducing confusion for people looking at waveforms.

The AHB spec does not require this and vendor VIP does not test for this.

Therefore, two changes were made to confirm this additional requirement is met:
- The AHBToTL block has an assertion that hrdata is steady during an error response
- The scoreboard fuzzer now checks that corrupt read data matches what was written

I have confirmed that both assertions fire when the hrdata register is removed.