GitHub Actions: Install the same packages in all jobs.
GitHub seems to have a very high failure rate when restoring artifacts
from the cache. When this happens, the build will attempt to compile
some of the cached artifacts, such as riscv-tools, from scratch. Because
we were not installing all of the packages required for building the
artifacts in the later jobs, if the cache restoration step fails, then
the build will fail.

Installing the same packages across all jobs will ensure that even if
the cache fails to restore, then our jobs will still succeed, albeit
more slowly.