Bump Verilator to 4.040, show seed on failure
Bump Verilator from 4.028 to 4.040 to pull in a change that adds
callbacks for when the simulation ends. This helps with debugging if
you have an intermittent failure and you hit a Chisel assertion.
Previously, you would have to be running verbosely to know what the
seed was. This change dumps the seed on exit so you always have a
reproducible seed ready to go.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>