Suppress an X-propagation in Rocket ITIM
ITIM writes are implemented as RMW operations and so cause ITIM reads.
Thus, initializing the ITIM can cause Xs to leak out.  But full-word
ITIM writes don't need to be RMW, so we can suppress the Xs from leaking
out in that special case.

Conveniently, we already require that TIMs be initialized with full-word
writes for ECC purposes, so preventing the Xs from leaking out only in the
full-word write case suffices.