IDecode: use DW_XPR for branch insts
The DW signal for branch insts was X,
and making it Bool(true) is fixing it
to one of the possible values.

No matter it is DW_X or DW_XPR, ALU would
function happily for RV32/RV64.

However, for RV64 ABLU, branch inst
depends on cmp, then xor, then adder_in1,
and if io.dw == 0, namely DW_32
and possibly DW_X (depending on decoder)
the higher 32 bits of adder_in1 is filled
with 1, then ABLU would malfunction.