config: add WithBootFreq configuration
The rocket-chip will set clock-frequency to 0 in the CPU fields of the
device tree. And the value is taken from bootFreqHz in the
RocketCoreParams which is zero by default. Some devices will use this
property to calculate the clock divider ratio, but a zero value will
cause division overflow.

This patch allows setting the value by simply adding WithBootFreq to the
configuration, so we will have the correct clock-frequency in the dts.

Signed-off-by: Yangyu Chen <cyy@cyyself.name>